Aplicación con buenas prácticas para el uso de API y Swagger

+++ src/main/java/es/televoip/controller/TaskControllerPostAndDelete.java
@@ -122,6 +122,10 @@ public class TaskControllerPostAndDelete {
    })
    @DeleteMapping("/{id}")
    public ResponseEntity<TaskDto> deleteTask(@Parameter(description = "id of task to be deleted") @PathVariable("id") Long id) {
+      if (service.findById(id) == null) {
+         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+      }
+
       service.deleteById(id);
       return new ResponseEntity<>(HttpStatus.OK);
    }
